[mod] refactor melli into package with submodules
